Refactoring:
* New ProclaimAPI class, moving API calls out of the main module instance class
* Move reference data out of main module instance class
