+++ refdata.js
@@ -1,0 +1,10 @@
+// Reference data used in setting up actions and presets
+export const song_parts = [
+	{ id: 0, label: 'Verse', path: 'verse' },
+	{ id: 1, label: 'Chorus', path: 'chorus' },
+	{ id: 2, label: 'Bridge', path: 'bridge' },
+	{ id: 3, label: 'Prechorus', displayLabel: 'Pre\nchorus', path: 'prechorus' },
+	{ id: 4, label: 'Interlude', displayLabel: 'Inter-lude', path: 'interlude' },
+	{ id: 5, label: 'Tag', path: 'tag' },
+	{ id: 6, label: 'Ending', path: 'ending' },
+]
